@@ -1,7 +1,7 @@
 package com.codephrase.android
 
-import android.databinding.BaseObservable
-import android.databinding.Bindable
+import androidx.databinding.BaseObservable
+import androidx.databinding.Bindable
 import com.codephrase.android.viewmodel.ViewModel
 
 private class BindingResources() : BaseObservable() {
