@@ -1,14 +1,14 @@
 package com.codephrase.android.activity
 
-import android.arch.lifecycle.Observer
 import android.os.Bundle
+import androidx.lifecycle.Observer
 import com.codephrase.android.common.navigation.NavigationHandler
 import com.codephrase.android.error.NotImplementedError
 import kotlin.reflect.KClass
 
 abstract class StartActivity : FrameActivity() {
     protected open val navigationTargetType: KClass<out FrameActivity>
-        get() = throw NotImplementedError()
+        get() = throw NotImplementedError("navigationTargetType")
 
     protected open val navigationHandler: NavigationHandler?
         get() = null
@@ -16,16 +16,16 @@ abstract class StartActivity : FrameActivity() {
     override fun onViewInitialized(savedInstanceState: Bundle?) {
         super.onViewInitialized(savedInstanceState)
 
-        viewModel.dataLoaded.observe(this, Observer { dataLoaded ->
-            if (dataLoaded == true) {
+        viewModel.dataLoaded.observe(this, Observer {
+            if (it == true) {
                 var type = navigationTargetType
                 var data: Any? = null
 
-                intent.data?.let { uri ->
-                    val navigationData = navigationHandler?.handleUri(uri)
-                    navigationData?.let { navigationData ->
-                        type = navigationData.type
-                        data = navigationData.data
+                intent.data?.let {
+                    val navigationData = navigationHandler?.handleUri(it)
+                    navigationData?.let {
+                        type = it.type
+                        data = it.data
                     }
                 }
 
