@@ -1,5 +1,5 @@
 package com.codephrase.android.error
 
-class NotImplementedError(message: String = "An operation is not implemented.") : Error(message) {
+class NotImplementedError(key: String) : Error("$key is not implemented.") {
 
 }
