@@ -1,5 +1,5 @@
 package com.codephrase.android.error
 
-class NotSupportedError(message: String = "An operation is not supported.") : Error(message) {
+class NotSupportedError(key: String) : Error("$key is not supported.") {
 
 }
