@@ -1,4 +1,4 @@
-package com.codephrase.android.error
+package com.codephrase.android.exception
 
 class NotFoundException(key: String) : Exception("$key doesn't exist.") {
 
