@@ -4,8 +4,8 @@ import android.content.Context
 import android.content.res.ColorStateList
 import android.content.res.TypedArray
 import android.graphics.drawable.Drawable
-import android.support.v4.content.ContextCompat
 import android.util.TypedValue
+import androidx.core.content.ContextCompat
 
 class ResourceHelper private constructor() {
     companion object {
